+++ Content/Core/Items/Accessories/Movement/Boots/StellarStreads.cs
@@ -1,32 +1,36 @@
+using System.Drawing;
+using Microsoft.Xna.Framework;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
 
 namespace TLR.Content.Core.Items.Accessories.Movement.Boots
-{ 
-    [AutoloadEquip(EquipType.Shoes)]
-	public class Shroomts : ModItem
+{
+	[AutoloadEquip(EquipType.Shoes)]
+	public class StellarStreads : ModItem
 	{
         public override void SetDefaults()
 		{
 			Item.width = 32;
 			Item.height = 28;
-			Item.value = Item.sellPrice(gold: 20);
-			Item.rare = ItemRarityID.Red;
+			Item.value = Item.sellPrice(gold: 25);
+			Item.rare = ItemRarityID.Expert;
+            Item.expert = true;
 			Item.accessory = true;
 			Item.ResearchUnlockCount = 1;
 			Item.maxStack = 1;
 		}
         public override void UpdateAccessory(Player player, bool hideVisual)
         {
-            player.accRunSpeed = 8f;
-            player.moveSpeed += 0.16f;
+            player.accRunSpeed = 9f;
+            player.moveSpeed += 0.20f;
             player.vanityRocketBoots = player.rocketBoots = 5;
             player.iceSkate = true;
             player.waterWalk = true;
             player.waterWalk2 = true;
             player.fireWalk = true;
-            player.lavaMax += 600;
+            player.lavaMax += 60;
+            player.lavaTime = player.lavaMax;
 			player.accFlipper = true;
 			player.accDivingHelm = true;
 			player.arcticDivingGear = true;
@@ -36,14 +40,15 @@ namespace TLR.Content.Core.Items.Accessories.Movement.Boots
 			player.frogLegJumpBoost = true;
 			player.noFallDmg = true;
 			player.autoJump = true;
+            player.empressBrooch = true;
         }
 
         public override void AddRecipes()
 		{
 			CreateRecipe()
-				.AddIngredient(ModContent.ItemType<MechaBoots>())
-				.AddIngredient(ItemID.MasterNinjaGear)
-				.AddIngredient(ItemID.ShroomiteBar, 18)
+				.AddIngredient(ModContent.ItemType<Shroomts>())
+				.AddIngredient(ItemID.EmpressFlightBooster)
+				.AddIngredient(ItemID.LunarBar, 10)
 				.AddTile(TileID.TinkerersWorkbench)
 				.Register();
 		}