+++ TLRItem.cs
@@ -1,7 +1,8 @@
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
-using TLR.Content.Core.Classes;
+using TLR.Content.Core.Items.Accessories.Combat.Defensive;
+using TLR.Content.Core.Tiles;
 
 namespace TLR
 {
@@ -9,15 +10,9 @@ namespace TLR
 	{
         public override void SetStaticDefaults()
         {
-            ItemID.Sets.ShimmerTransformToItem[ItemID.CrystalBall] = ModContent.ItemType<Content.Core.Tiles.ShimmerBallItem>();
-            ItemID.Sets.ShimmerTransformToItem[ItemID.CobaltShield] = ModContent.ItemType<Content.Core.Items.Accessories.Combat.Defensive.PalladiumShield>();
-            ItemID.Sets.ShimmerTransformToItem[ModContent.ItemType<Content.Core.Items.Accessories.Combat.Defensive.PalladiumShield>()] = ItemID.CobaltShield;
-        }
-        public override void SetDefaults(Item entity)
-        {
-            if (entity.type == ItemID.Flymeal && ModContent.GetInstance<TLRConfigServer>().ItemReclassing) {
-                entity.DamageType = ModContent.GetInstance<SadistMelee>();
-            }
+            ItemID.Sets.ShimmerTransformToItem[ItemID.CrystalBall] = ModContent.ItemType<ShimmerBallItem>();
+            ItemID.Sets.ShimmerTransformToItem[ItemID.CobaltShield] = ModContent.ItemType<PalladiumShield>();
+            ItemID.Sets.ShimmerTransformToItem[ModContent.ItemType<PalladiumShield>()] = ItemID.CobaltShield;
         }
     }
 }