+++ TLRPlayer.cs
@@ -31,13 +31,9 @@ namespace TLR
             if (brokenHeart) { Player.statLife += 10; }
             if (item.DamageType.CountsAsClass(DamageClass.Melee))
             {
-                if (hallowGlove && hit.Crit) {
-                    if ((target.life / target.lifeMax) <= 0.15) {
-                        target.life = 0;
-                    }
-                    else {
-                        target.AddBuff(ModContent.BuffType<BrokenBlessing>(), 600, false);
-                    }
+                if (hallowGlove) {
+                    if (hit.Crit) { target.AddBuff(BuffID.Frozen, 180); }
+                    target.AddBuff(BuffID.Frostburn2, 300);
                 }
             }
         }
@@ -46,13 +42,9 @@ namespace TLR
             if (brokenHeart) { Player.statLife += 10; }
             if (proj.DamageType.CountsAsClass(DamageClass.Melee))
             {
-                if (hallowGlove && hit.Crit) {
-                    if ((target.life / target.lifeMax) <= 0.1) {
-                        target.life = 0;
-                    }
-                    else {
-                        target.AddBuff(ModContent.BuffType<BrokenBlessing>(), 300, false);
-                    }
+                if (hallowGlove) {
+                    if (hit.Crit) { target.AddBuff(BuffID.Frozen, 180); }
+                    target.AddBuff(BuffID.Frostburn2, 300);
                 }
             }
         }